LayoutTests/fast/js/script-line-number-expected.txt

This test checks that line numbers are correctly reported for both inline scripts and inline event handlers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 14 is 14
PASS 17 is 17
PASS successfullyParsed is true

TEST COMPLETE
